refactor(eslint-config): group pure-data modules under lib/
Move `constants.js` to `lib/file-patterns.js` and `browser-globals.js` to
`lib/browser-globals.js` to separate pure-data modules from config-output
modules in `src/`. No change to exports or behaviour.

Closes #72
