+++ packages/eslint-config/src/lib/browser-globals.js
