+++ packages/eslint-config/src/lib/file-patterns.js
