+++ packages/eslint-config/src/plugins/compat.js
@@ -1,6 +1,6 @@
 import eslintPluginCompat from 'eslint-plugin-compat';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
