+++ packages/eslint-config/src/plugins/depend.js
@@ -1,6 +1,6 @@
 import eslintPluginDepend from 'eslint-plugin-depend';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
