+++ packages/eslint-config/src/plugins/import.js
@@ -7,7 +7,7 @@ import {
   JS_EXTENSIONS,
   JS_FILES,
   TEST_FILES,
-} from '../constants.js';
+} from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
