+++ packages/eslint-config/src/plugins/n.js
@@ -1,6 +1,6 @@
 import eslintPluginN from 'eslint-plugin-n';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
