+++ packages/eslint-config/src/plugins/next.js
@@ -1,6 +1,6 @@
 import eslintPluginNext from '@next/eslint-plugin-next';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
