+++ packages/eslint-config/src/plugins/no-unsanitized.js
@@ -1,6 +1,6 @@
 import eslintPluginNoUnsanitized from 'eslint-plugin-no-unsanitized';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
