+++ packages/eslint-config/src/plugins/no-use-extend-native.js
@@ -1,6 +1,6 @@
 import eslintPluginNoUseExtendNative from 'eslint-plugin-no-use-extend-native';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
