+++ packages/eslint-config/src/plugins/promise.js
@@ -1,6 +1,6 @@
 import eslintPluginPromise from 'eslint-plugin-promise';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
