+++ packages/eslint-config/src/plugins/sonarjs.js
@@ -1,6 +1,6 @@
 import eslintPluginSonarjs from 'eslint-plugin-sonarjs';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
