+++ packages/eslint-config/src/plugins/tailwindcss.js
@@ -1,6 +1,6 @@
 import eslintPluginBetterTailwindcss from 'eslint-plugin-better-tailwindcss';
 
-import { DEFAULT_FILES } from '../constants.js';
+import { DEFAULT_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
