+++ packages/eslint-config/src/plugins/vitest.js
@@ -1,6 +1,6 @@
 import eslintPluginVitest from '@vitest/eslint-plugin';
 
-import { TEST_FILES } from '../constants.js';
+import { TEST_FILES } from '../lib/file-patterns.js';
 
 /** @import { Linter } from 'eslint' */
